Re-essentialisation
Added some corrections as described in pull request, still a lot to come before a merge
