+++ .pep8speaks.yml
@@ -12,6 +12,7 @@ pycodestyle:  # Same as scanner.linter value. Other option is flake8
         - E731  # do not assign a lambda expression, use a def
         - C406  # Unnecessary list literal - rewrite as a dict literal.
         - E741  # ambiguous variable name
+        - W503  # line break before binary operator
 
 no_blank_comment: False  # If True, no comment is made on PR without any errors.
 descending_issues_order: False  # If True, PEP 8 issues in message will be displayed in descending order of line numbers in the file
