+++ NotIncluded/NotIncludedYet/ClassesGoodwin_Depreciated.py
