+++ NotIncluded/NotIncludedYet/Miscfunc.py
