+++ NotIncluded/NotIncludedYet/SpatialOperators.py
