+++ NotIncluded/NotIncludedYet/_Check_SystemOfEquation.py
