+++ NotIncluded/SystemOfEquations/GK_Noinflation.py
