+++ NotIncluded/SystemOfEquations/G_Reduced.py
